+++ aes/Cargo.toml
@@ -15,7 +15,7 @@ categories = ["cryptography", "no-std"]
 [dependencies]
 cipher = "0.5"
 cpubits = "0.1"
-zeroize = { version = "1.8", optional = true, default-features = false, features = ["aarch64"] }
+zeroize = { version = "1.9", optional = true, default-features = false }
 
 [target.'cfg(any(target_arch = "aarch64", target_arch = "x86_64", target_arch = "x86"))'.dependencies]
 cpufeatures = "0.3"
