+++ aes/src/backends/aarch64_aes/expand.rs
@@ -3,9 +3,7 @@
 
 use core::{arch::aarch64::*, mem, slice};
 
-pub(super) type Aes128RoundKeys = [uint8x16_t; 11];
-pub(super) type Aes192RoundKeys = [uint8x16_t; 13];
-pub(super) type Aes256RoundKeys = [uint8x16_t; 15];
+pub(super) type RoundKeys<const RK: usize> = [uint8x16_t; RK];
 
 /// There are 4 AES words in a block.
 const BLOCK_WORDS: usize = 4;
@@ -17,16 +15,17 @@ const WORD_SIZE: usize = 4;
 const ROUND_CONSTS: [u32; 10] = [0x01, 0x02, 0x04, 0x08, 0x10, 0x20, 0x40, 0x80, 0x1b, 0x36];
 
 /// AES key expansion.
+#[inline]
 #[target_feature(enable = "aes")]
-pub unsafe fn expand_key<const L: usize, const N: usize>(key: &[u8; L]) -> [uint8x16_t; N] {
-    assert!((L == 16 && N == 11) || (L == 24 && N == 13) || (L == 32 && N == 15));
+pub unsafe fn expand_key<const N: usize, const RK: usize>(key: &[u8; N]) -> [uint8x16_t; RK] {
+    const { assert!(matches!((N, RK), (16, 11) | (24, 13) | (32, 15))) }
 
-    let mut expanded_keys: [uint8x16_t; N] = mem::zeroed();
+    let mut expanded_keys: [uint8x16_t; RK] = mem::zeroed();
 
     // Sanity check, as this is required in order for the subsequent conversion to be sound.
     const _: () = assert!(mem::align_of::<uint8x16_t>() >= mem::align_of::<u32>());
     let keys_ptr: *mut u32 = expanded_keys.as_mut_ptr().cast();
-    let columns = slice::from_raw_parts_mut(keys_ptr, N * BLOCK_WORDS);
+    let columns = slice::from_raw_parts_mut(keys_ptr, RK * BLOCK_WORDS);
 
     for (i, chunk) in key.chunks_exact(WORD_SIZE).enumerate() {
         columns[i] = u32::from_ne_bytes(chunk.try_into().unwrap());
@@ -35,9 +34,9 @@ pub unsafe fn expand_key<const L: usize, const N: usize>(key: &[u8; L]) -> [uint
     // From "The Rijndael Block Cipher" Section 4.1:
     // > The number of columns of the Cipher Key is denoted by `Nk` and is
     // > equal to the key length divided by 32 [bits].
-    let nk = L / WORD_SIZE;
+    let nk = N / WORD_SIZE;
 
-    for i in nk..(N * BLOCK_WORDS) {
+    for i in nk..(RK * BLOCK_WORDS) {
         let mut word = columns[i - 1];
 
         if i % nk == 0 {
@@ -56,16 +55,19 @@ pub unsafe fn expand_key<const L: usize, const N: usize>(key: &[u8; L]) -> [uint
 ///
 /// This is the reverse of the encryption keys, with the Inverse Mix Columns
 /// operation applied to all but the first and last expanded key.
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn inv_expanded_keys<const N: usize>(keys: &[uint8x16_t; N]) -> [uint8x16_t; N] {
-    assert!(N == 11 || N == 13 || N == 15);
+pub(super) unsafe fn inv_expanded_keys<const RK: usize>(
+    keys: &[uint8x16_t; RK],
+) -> [uint8x16_t; RK] {
+    const { assert!(matches!(RK, 11 | 13 | 15)) }
 
-    let mut inv_keys: [uint8x16_t; N] = core::mem::zeroed();
-    inv_keys[0] = keys[N - 1];
-    for i in 1..N - 1 {
-        inv_keys[i] = vaesimcq_u8(keys[N - 1 - i]);
+    let mut inv_keys: [uint8x16_t; RK] = core::mem::zeroed();
+    inv_keys[0] = keys[RK - 1];
+    for i in 1..RK - 1 {
+        inv_keys[i] = vaesimcq_u8(keys[RK - 1 - i]);
     }
-    inv_keys[N - 1] = keys[0];
+    inv_keys[RK - 1] = keys[0];
 
     inv_keys
 }
