+++ aes/src/backends/x86_aes/encdec.rs
@@ -1,55 +1,56 @@
 #![allow(unsafe_op_in_unsafe_fn)]
 
-use super::arch::*;
+use super::RoundKeys;
 use crate::Block;
 use cipher::{
     array::{Array, ArraySize},
     inout::InOut,
 };
 
+#[cfg(target_arch = "x86")]
+use core::arch::x86::*;
+#[cfg(target_arch = "x86_64")]
+use core::arch::x86_64::*;
+
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn encrypt<const KEYS: usize>(
-    keys: &[__m128i; KEYS],
-    block: InOut<'_, '_, Block>,
-) {
-    assert!(KEYS == 11 || KEYS == 13 || KEYS == 15);
+pub(crate) unsafe fn encrypt<const RK: usize>(keys: &RoundKeys<RK>, block: InOut<'_, '_, Block>) {
+    const { assert!(matches!(RK, 11 | 13 | 15)) }
 
     let (block_in, block_out) = block.into_raw();
     let mut b = _mm_loadu_si128(block_in.cast());
     b = _mm_xor_si128(b, keys[0]);
-    for &key in &keys[1..KEYS - 1] {
+    for &key in &keys[1..RK - 1] {
         b = _mm_aesenc_si128(b, key);
     }
-    b = _mm_aesenclast_si128(b, keys[KEYS - 1]);
+    b = _mm_aesenclast_si128(b, keys[RK - 1]);
     _mm_storeu_si128(block_out.cast(), b);
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn decrypt<const KEYS: usize>(
-    keys: &[__m128i; KEYS],
-    block: InOut<'_, '_, Block>,
-) {
-    assert!(KEYS == 11 || KEYS == 13 || KEYS == 15);
+pub(crate) unsafe fn decrypt<const RK: usize>(keys: &RoundKeys<RK>, block: InOut<'_, '_, Block>) {
+    const { assert!(matches!(RK, 11 | 13 | 15)) }
 
     let (block_in, block_out) = block.into_raw();
     let mut b = _mm_loadu_si128(block_in.cast());
     b = _mm_xor_si128(b, keys[0]);
-    for &key in &keys[1..KEYS - 1] {
+    for &key in &keys[1..RK - 1] {
         b = _mm_aesdec_si128(b, key);
     }
-    b = _mm_aesdeclast_si128(b, keys[KEYS - 1]);
+    b = _mm_aesdeclast_si128(b, keys[RK - 1]);
     _mm_storeu_si128(block_out.cast(), b);
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn encrypt_par<const KEYS: usize, ParBlocks: ArraySize>(
-    keys: &[__m128i; KEYS],
-    blocks: InOut<'_, '_, Array<Block, ParBlocks>>,
+pub(super) unsafe fn encrypt_par<const RK: usize, ParBlocks: ArraySize>(
+    keys: &RoundKeys<RK>,
+    mut blocks: InOut<'_, '_, Array<Block, ParBlocks>>,
 ) {
-    assert!(KEYS == 11 || KEYS == 13 || KEYS == 15);
+    const { assert!(matches!(RK, 11 | 13 | 15)) }
 
-    let (blocks_in, blocks_out) = blocks.into_raw();
-    let mut b = load(blocks_in);
+    let mut b = load(blocks.get_in());
 
     // Loop over keys is intentionally not used here to force inlining
     xor(&mut b, keys[0]);
@@ -62,27 +63,27 @@ pub(super) unsafe fn encrypt_par<const KEYS: usize, ParBlocks: ArraySize>(
     aesenc(&mut b, keys[7]);
     aesenc(&mut b, keys[8]);
     aesenc(&mut b, keys[9]);
-    if KEYS >= 13 {
+    if RK >= 13 {
         aesenc(&mut b, keys[10]);
         aesenc(&mut b, keys[11]);
     }
-    if KEYS == 15 {
+    if RK == 15 {
         aesenc(&mut b, keys[12]);
         aesenc(&mut b, keys[13]);
     }
-    aesenclast(&mut b, keys[KEYS - 1]);
-    store(blocks_out, b);
+    aesenclast(&mut b, keys[RK - 1]);
+    store(blocks.get_out(), b);
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn decrypt_par<const KEYS: usize, ParBlocks: ArraySize>(
-    keys: &[__m128i; KEYS],
-    blocks: InOut<'_, '_, Array<Block, ParBlocks>>,
+pub(super) unsafe fn decrypt_par<const RK: usize, ParBlocks: ArraySize>(
+    keys: &RoundKeys<RK>,
+    mut blocks: InOut<'_, '_, Array<Block, ParBlocks>>,
 ) {
-    assert!(KEYS == 11 || KEYS == 13 || KEYS == 15);
+    const { assert!(matches!(RK, 11 | 13 | 15)) };
 
-    let (blocks_in, blocks_out) = blocks.into_raw();
-    let mut b = load(blocks_in);
+    let mut b = load(blocks.get_in());
 
     // Loop over keys is intentionally not used here to force inlining
     xor(&mut b, keys[0]);
@@ -95,33 +96,29 @@ pub(super) unsafe fn decrypt_par<const KEYS: usize, ParBlocks: ArraySize>(
     aesdec(&mut b, keys[7]);
     aesdec(&mut b, keys[8]);
     aesdec(&mut b, keys[9]);
-    if KEYS >= 13 {
+    if RK >= 13 {
         aesdec(&mut b, keys[10]);
         aesdec(&mut b, keys[11]);
     }
-    if KEYS == 15 {
+    if RK == 15 {
         aesdec(&mut b, keys[12]);
         aesdec(&mut b, keys[13]);
     }
-    aesdeclast(&mut b, keys[KEYS - 1]);
-    store(blocks_out, b);
+    aesdeclast(&mut b, keys[RK - 1]);
+    store(blocks.get_out(), b);
 }
 
 #[target_feature(enable = "sse2")]
-pub(crate) unsafe fn load<N: ArraySize>(blocks: *const Array<Block, N>) -> Array<__m128i, N> {
-    let p = blocks.cast::<__m128i>();
-    let mut res: Array<__m128i, N> = core::mem::zeroed();
-    for i in 0..N::USIZE {
-        res[i] = _mm_loadu_si128(p.add(i));
-    }
-    res
+pub(crate) unsafe fn load<N: ArraySize>(blocks: &Array<Block, N>) -> Array<__m128i, N> {
+    let p: *const __m128i = blocks.as_ptr().cast();
+    Array::from_fn(|i| unsafe { _mm_loadu_si128(p.add(i)) })
 }
 
 #[target_feature(enable = "sse2")]
-pub(crate) unsafe fn store<N: ArraySize>(blocks: *mut Array<Block, N>, b: Array<__m128i, N>) {
-    let p = blocks.cast::<__m128i>();
-    for i in 0..N::USIZE {
-        _mm_storeu_si128(p.add(i), b[i]);
+pub(crate) unsafe fn store<N: ArraySize>(dst: &mut Array<Block, N>, blocks: Array<__m128i, N>) {
+    let p: *mut __m128i = dst.as_mut_ptr().cast();
+    for (i, block) in blocks.into_iter().enumerate() {
+        unsafe { _mm_storeu_si128(p.add(i), block) }
     }
 }
