+++ aes/src/backends/x86_aes/expand.rs
@@ -1,20 +1,21 @@
 #![allow(unsafe_op_in_unsafe_fn)]
 
-use super::arch::*;
-use core::mem::{transmute, zeroed};
+#[cfg(target_arch = "x86")]
+use core::arch::x86::*;
+#[cfg(target_arch = "x86_64")]
+use core::arch::x86_64::*;
 
-pub(super) type Aes128RoundKeys = [__m128i; 11];
-pub(super) type Aes192RoundKeys = [__m128i; 13];
-pub(super) type Aes256RoundKeys = [__m128i; 15];
+pub(crate) type RoundKeys<const RK: usize> = [__m128i; RK];
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn aes128_expand_key(key: &[u8; 16]) -> Aes128RoundKeys {
-    unsafe fn expand_round<const RK: i32>(keys: &mut Aes128RoundKeys, pos: usize) {
+pub(super) unsafe fn aes128_expand_key(key: &[u8; 16]) -> RoundKeys<11> {
+    unsafe fn expand_round<const R: i32>(keys: &mut RoundKeys<11>, pos: usize) {
         let mut t1 = keys[pos - 1];
         let mut t2;
         let mut t3;
 
-        t2 = _mm_aeskeygenassist_si128(t1, RK);
+        t2 = _mm_aeskeygenassist_si128(t1, R);
         t2 = _mm_shuffle_epi32(t2, 0xff);
         t3 = _mm_slli_si128(t1, 0x4);
         t1 = _mm_xor_si128(t1, t3);
@@ -27,7 +28,7 @@ pub(super) unsafe fn aes128_expand_key(key: &[u8; 16]) -> Aes128RoundKeys {
         keys[pos] = t1;
     }
 
-    let mut keys: Aes128RoundKeys = zeroed();
+    let mut keys = [_mm_setzero_si128(); 11];
     let k = _mm_loadu_si128(key.as_ptr().cast());
     keys[0] = k;
 
@@ -46,19 +47,19 @@ pub(super) unsafe fn aes128_expand_key(key: &[u8; 16]) -> Aes128RoundKeys {
     keys
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> Aes192RoundKeys {
-    unsafe fn shuffle(a: __m128i, b: __m128i, i: usize) -> __m128i {
-        let a: [u64; 2] = transmute(a);
-        let b: [u64; 2] = transmute(b);
-        transmute([a[i], b[0]])
+pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> RoundKeys<13> {
+    unsafe fn unpack_hilo(a: __m128i, b: __m128i) -> __m128i {
+        let a = _mm_shuffle_epi32(a, 0b01_00_11_10);
+        _mm_unpacklo_epi64(a, b)
     }
 
     #[target_feature(enable = "aes")]
-    unsafe fn expand_round<const RK: i32>(mut t1: __m128i, mut t3: __m128i) -> (__m128i, __m128i) {
+    unsafe fn expand_round<const R: i32>(mut t1: __m128i, mut t3: __m128i) -> (__m128i, __m128i) {
         let (mut t2, mut t4);
 
-        t2 = _mm_aeskeygenassist_si128(t3, RK);
+        t2 = _mm_aeskeygenassist_si128(t3, R);
         t2 = _mm_shuffle_epi32(t2, 0x55);
         t4 = _mm_slli_si128(t1, 0x4);
         t1 = _mm_xor_si128(t1, t4);
@@ -75,7 +76,7 @@ pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> Aes192RoundKeys {
         (t1, t3)
     }
 
-    let mut keys: Aes192RoundKeys = zeroed();
+    let mut keys = [_mm_setzero_si128(); 13];
     // We are being extra pedantic here to remove out-of-bound access.
     // This should be optimized into movups, movsd sequence.
     let (k0, k1l) = {
@@ -90,15 +91,15 @@ pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> Aes192RoundKeys {
     keys[0] = k0;
 
     let (k1_2, k2r) = expand_round::<0x01>(k0, k1l);
-    keys[1] = shuffle(k1l, k1_2, 0);
-    keys[2] = shuffle(k1_2, k2r, 1);
+    keys[1] = _mm_unpacklo_epi64(k1l, k1_2);
+    keys[2] = unpack_hilo(k1_2, k2r);
 
     let (k3, k4l) = expand_round::<0x02>(k1_2, k2r);
     keys[3] = k3;
 
     let (k4_5, k5r) = expand_round::<0x04>(k3, k4l);
-    let k4 = shuffle(k4l, k4_5, 0);
-    let k5 = shuffle(k4_5, k5r, 1);
+    let k4 = _mm_unpacklo_epi64(k4l, k4_5);
+    let k5 = unpack_hilo(k4_5, k5r);
     keys[4] = k4;
     keys[5] = k5;
 
@@ -106,15 +107,15 @@ pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> Aes192RoundKeys {
     keys[6] = k6;
 
     let (k7_8, k8r) = expand_round::<0x10>(k6, k7l);
-    keys[7] = shuffle(k7l, k7_8, 0);
-    keys[8] = shuffle(k7_8, k8r, 1);
+    keys[7] = _mm_unpacklo_epi64(k7l, k7_8);
+    keys[8] = unpack_hilo(k7_8, k8r);
 
     let (k9, k10l) = expand_round::<0x20>(k7_8, k8r);
     keys[9] = k9;
 
     let (k10_11, k11r) = expand_round::<0x40>(k9, k10l);
-    keys[10] = shuffle(k10l, k10_11, 0);
-    keys[11] = shuffle(k10_11, k11r, 1);
+    keys[10] = _mm_unpacklo_epi64(k10l, k10_11);
+    keys[11] = unpack_hilo(k10_11, k11r);
 
     let (k12, _) = expand_round::<0x80>(k10_11, k11r);
     keys[12] = k12;
@@ -122,15 +123,16 @@ pub(super) unsafe fn aes192_expand_key(key: &[u8; 24]) -> Aes192RoundKeys {
     keys
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn aes256_expand_key(key: &[u8; 32]) -> Aes256RoundKeys {
-    unsafe fn expand_round<const RK: i32>(keys: &mut Aes256RoundKeys, pos: usize) {
+pub(super) unsafe fn aes256_expand_key(key: &[u8; 32]) -> RoundKeys<15> {
+    unsafe fn expand_round<const R: i32>(keys: &mut RoundKeys<15>, pos: usize) {
         let mut t1 = keys[pos - 2];
         let mut t2;
         let mut t3 = keys[pos - 1];
         let mut t4;
 
-        t2 = _mm_aeskeygenassist_si128(t3, RK);
+        t2 = _mm_aeskeygenassist_si128(t3, R);
         t2 = _mm_shuffle_epi32(t2, 0xff);
         t4 = _mm_slli_si128(t1, 0x4);
         t1 = _mm_xor_si128(t1, t4);
@@ -155,13 +157,13 @@ pub(super) unsafe fn aes256_expand_key(key: &[u8; 32]) -> Aes256RoundKeys {
         keys[pos + 1] = t3;
     }
 
-    unsafe fn expand_round_last<const RK: i32>(keys: &mut Aes256RoundKeys, pos: usize) {
+    unsafe fn expand_round_last<const R: i32>(keys: &mut RoundKeys<15>, pos: usize) {
         let mut t1 = keys[pos - 2];
         let mut t2;
         let t3 = keys[pos - 1];
         let mut t4;
 
-        t2 = _mm_aeskeygenassist_si128(t3, RK);
+        t2 = _mm_aeskeygenassist_si128(t3, R);
         t2 = _mm_shuffle_epi32(t2, 0xff);
         t4 = _mm_slli_si128(t1, 0x4);
         t1 = _mm_xor_si128(t1, t4);
@@ -174,7 +176,7 @@ pub(super) unsafe fn aes256_expand_key(key: &[u8; 32]) -> Aes256RoundKeys {
         keys[pos] = t1;
     }
 
-    let mut keys: Aes256RoundKeys = zeroed();
+    let mut keys = [_mm_setzero_si128(); 15];
 
     let kp = key.as_ptr().cast::<__m128i>();
     keys[0] = _mm_loadu_si128(kp);
@@ -192,9 +194,10 @@ pub(super) unsafe fn aes256_expand_key(key: &[u8; 32]) -> Aes256RoundKeys {
     keys
 }
 
+#[inline]
 #[target_feature(enable = "aes")]
-pub(super) unsafe fn inv_keys<const N: usize>(keys: &[__m128i; N]) -> [__m128i; N] {
-    let mut inv_keys: [__m128i; N] = zeroed();
+pub(super) unsafe fn inv_expanded_keys<const N: usize>(keys: &[__m128i; N]) -> [__m128i; N] {
+    let mut inv_keys: [__m128i; N] = [_mm_setzero_si128(); N];
     inv_keys[0] = keys[N - 1];
     for i in 1..N - 1 {
         inv_keys[i] = _mm_aesimc_si128(keys[N - 1 - i]);
