--- aes/src/x86/vaes256.rs
@@ -1,1 +1,0 @@
-pub(super) mod encdec;
