--- aes/src/x86/vaes512.rs
@@ -1,1 +1,0 @@
-pub(super) mod encdec;
